feat: migrate book controller

+++ day-5-6/internal/services/book.go
@@ -4,13 +4,14 @@ import (
 	"alterra-agmc-day-5-6/internal/models"
 	"alterra-agmc-day-5-6/internal/repositories"
 	"context"
+	"errors"
 )
 
 type BookService interface {
 	FindAll(ctx context.Context) ([]*models.Book, error)
 	FindByID(ctx context.Context, id uint) (*models.Book, error)
 	Create(ctx context.Context, book *models.Book) (*models.Book, error)
-	DeleteByID(ctx context.Context, id uint) error
+	DeleteByID(ctx context.Context, id uint, userId uint) error
 	Update(ctx context.Context, book *models.Book) (*models.Book, error)
 }
 
@@ -24,7 +25,14 @@ func (s *bookServiceImpl) Create(ctx context.Context, book *models.Book) (*model
 }
 
 // DeleteByID implements BookService
-func (s *bookServiceImpl) DeleteByID(ctx context.Context, id uint) error {
+func (s *bookServiceImpl) DeleteByID(ctx context.Context, id uint, userID uint) error {
+	book, err := s.FindByID(ctx, id)
+	if err != nil {
+		return err
+	}
+	if book.UserID != userID {
+		return errors.New("access denied")
+	}
 	return s.repo.DeleteByID(ctx, id)
 }
 
@@ -40,6 +48,13 @@ func (s *bookServiceImpl) FindByID(ctx context.Context, id uint) (*models.Book, 
 
 // Update implements BookService
 func (s *bookServiceImpl) Update(ctx context.Context, book *models.Book) (*models.Book, error) {
+	b, err := s.FindByID(ctx, book.ID)
+	if err != nil {
+		return nil, err
+	}
+	if b.UserID != book.UserID {
+		return nil, errors.New("access denied")
+	}
 	return s.repo.Update(ctx, book)
 }
